requirements.txt
flask==2.0.1
werkzeug==2.0.1
flask-cors==3.0.10
pyzbar==0.1.8
pillow==8.3.1
opencv-python-headless==4.5.3.56
numpy==1.21.1